Moved login mehtods to Login class
